requirements.txt
certifi>=2020.12.5
chardet>=4.0.0
click>=7.1.2
python-crontab>=2.5.1
Flask>=1.1.2
idna>=2.10
itsdangerous>=1.1.0
Jinja2>=2.11.3
MarkupSafe>=1.1.1
PyMySQL>=1.0.2
requests>=2.25.1
urllib3>=1.26.3
Werkzeug>=1.0.1
consulate>=0.6.0
prometheus-flask-exporter>=0.18.3